Trying to fix limitation to integer fields

+++ livebot/models.py
@@ -80,7 +80,7 @@ class TwitchNotification(models.Model):
     """
     twitch = models.ForeignKey(TwitchChannel, verbose_name='Twitch Channel')
     content_type = models.ForeignKey(ContentType, on_delete=models.CASCADE)
-    object_id = models.PositiveIntegerField()
+    object_id = models.BigIntegerField()
     content_object = GenericForeignKey('content_type', 'object_id')
     message = models.CharField(max_length=255, default='{name} is Live!', verbose_name='Notification Message')
 
@@ -129,7 +129,7 @@ class Notification(models.Model):
     log = models.ForeignKey('Log', verbose_name='Log Item')
     live = models.ForeignKey(TwitchLive, verbose_name='Twitch Live')
     content_type = models.ForeignKey(ContentType, on_delete=models.CASCADE)
-    object_id = models.PositiveIntegerField()
+    object_id = models.BigIntegerField()
     content_object = GenericForeignKey('content_type', 'object_id')
     success = models.BooleanField(default=False, verbose_name='Success')
